third_party/blink/web_tests/external/wpt/css/css-inline/parsing/alignment-baseline-invalid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['alignment-baseline'] = "auto" should not set the property value
  assert_equals: expected "" but got "auto"
Harness: the test ran to completion.
